Modernize native interop for NET7+ by switching from DllImport to LibraryImport.

+++ WmiLight/Wbem/WbemLocator.cs
@@ -89,7 +89,7 @@
 
             IntPtr wbemServices;
 
-            HResult hResult = NativeMethods.ConnectServer(this, networkResource, userName, userPassword, locale, wbemConnectOption, authority, ctx, out wbemServices);
+            HResult hResult = NativeMethods.ConnectServer(this, networkResource, userName, userPassword, locale, (uint)wbemConnectOption, authority, ctx, out wbemServices);
 
             if (hResult.Failed)
             {
